Refactor typed OMML translator internals
